vulnerable_app/requirements.txt
Flask>=2.0
Flask-CORS>=3.0
streamlit>=1.0
SQLAlchemy>=1.4
PyMySQL>=1.0
cryptography>=46.0
python-dotenv>=0.19
requests
